Created tag colossus1_15.
git-svn-id: file:///cygdrive/c/CTTC/backup/redyc/svn/Colossus/tags/colossus1_15@768 8e1883df-7ee1-ef4f-9e60-6d04f2a2cf0e

+++ src/extensiones/udb/bdd.c
@@ -156,7 +156,7 @@ u_long ObtieneHash(UDBloq *bloq)
 	if (fstat(bloq->fd, &inode) < 0)
 		return 0L;
 	t = inode.st_size - INI_DATA;
-	if (t <= 0)
+	if ((int)t <= 0)
 		return 0L;
 	par = Malloc(t);
 	lseek(bloq->fd, INI_DATA, SEEK_SET);
